Refactored and modularized, focusing on folder organization for styles and components
