+++ Components/Layout.jsx
@@ -1,6 +1,4 @@
-// components/layout.js
-
-import Header from '../Components/Header'
+import Header from '../Components/Header';
 
 export default function Layout({ children }) {
   return (
