--- styles/Home.module.css
@@ -1,437 +1,0 @@
-.container {
-  padding: 0 0rem;
-}
-
-.homeLink {
-  color: inherit;
-}
-
-.breathe_container {
-  text-align: center;
-}
-
-.title_circle {
-  position: relative;
-  fill: black;
-  transform: scale(0);
-  animation: title_circle_load, title_circle_heartbeat;
-  animation-duration: 0.5s, 10s;
-  animation-delay: 1s, 4s;
-  animation-iteration-count: 1, infinite;
-  animation-play-state: running, running;
-  animation-fill-mode: forwards, forwards;
-  z-index: 2;
-}
-
-@keyframes title_circle_load {
-  0% {
-   transform: scale(0);
-   fill: black;
-  }
-  100% {
-   transform: scale(0.12);
-   fill: black;
-  }
-}
-
-@keyframes title_circle_heartbeat {
-  0% {
-   transform: scale(0.12);
-   filter: invert(0%);
-  }
-  50% {
-   transform: scale(2);
-   filter: invert(100%);
-  }
-  100% {
-   transform: scale(0.12);
-   filter: invert(0%);
-  }
-}
-
-.breathe_instructions_1 {
-  position: relative;
-  cursor: default;
-  font-family: Inter;
-  opacity: 0;
-  color: white;
-  animation: breath_instructions_fade_in, breath_instructions_fade_out;
-  animation-duration: 0.5s, 0.5s;
-  animation-delay: 0s, 8s;
-  animation-iteration-count: 1;
-  animation-fill-mode: forwards;
-  z-index: 1;
-  transform: translateY(-50px);
-}
-
-.breathe_instructions_2 {
-  position: relative;
-  cursor: default;
-  font-family: Inter;
-  opacity: 0;
-  color: white;
-  animation: breath_instructions_fade_in, breath_instructions_fade_out;
-  animation-duration: 0.5s, 0.5s;
-  animation-delay: 19s, 29s;
-  animation-iteration-count: 1;
-  animation-fill-mode: forwards;
-  z-index: 1;
-  transform: translateY(-50px);
-}
-
-@keyframes breath_instructions_fade_in {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-@keyframes breath_instructions_fade_out {
-  0% {
-    opacity: 1;
-  }
-  100% {
-    opacity: 0;
-  }
-}
- 
-.main {
-  min-height: 100vh;
-  width: 100%;
-  padding: 0;
-  margin: 0;
-  flex: 1;
-  position: absolute;
-  top: 0;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-}
-
-.footer {
-  display: flex;
-  flex: 1;
-  padding: 2rem 0;
-  border-top: 1px solid #eaeaea;
-  justify-content: center;
-  align-items: center;
-}
-
-.homeLink {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  flex-grow: 1;
-}
-
-.title a {
-  color: inherit;
-  text-decoration: none;
-}
-
-.title a:hover,
-.title a:focus,
-.title a:active {
-  text-decoration: inherit;
-}
-
-.title {
-  margin: 0;
-  line-height: 2rem;
-  font-size: 2rem;
-  align-self: center;
-  text-orientation: upright;
-  color: #ffffff;
-}
-
-.subtitle {
-  margin: 0;
-  line-height: 0.5rem;
-  font-size: 1rem;
-  align-self: flex-start;
-  text-orientation: mixed;
-}
-
-.description {
-  margin: 0 0;
-  line-height: 0.2rem;
-  font-size: 0.4rem;
-  text-orientation: mixed;
-  align-self: center;
-  font-style: italic;
-}
-
-.title_container {
-  padding: 0;
-  margin: 0;
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  position: fixed;
-  top: 10px;
-  left: 10px;
-  z-index: 99;
-  cursor: pointer;
-}
-
-.title,
-.subtitle,
-.description {
-  text-align: center;
-}
-
-.section_block {
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  min-width: 33rem;
-}
-
-.section_title {
-  font-family: Ovo-Regular;
-  color: black;
-  font-size: 5rem;
-  line-height: 5rem;
-  font-weight: normal;
-  margin: 0;
-}
-
-.section_paragraph {
-  font-family: Ovo-Regular;
-  color: black;
-  font-size: 1.5rem;
-  line-height: 1.5rem;
-  font-weight: normal;
-  align-self: flex-end;
-  margin: 0;
-}
-
-.logo_block {
-  display: flex;
-  flex-direction: row;
-  z-index: 99;
-}
-
-.logo_block img {
-  margin-top: 5px;
-  margin-right: 5px;
-  background-color: white;
-}
-
-.character_tree {
-  opacity: 1;
-  animation: graph_shift;
-  animation-duration: 5s;
-  animation-timing-function: ease-in-out;
-  animation-delay: 10s;
-  animation-fill-mode: forwards;
-}
-
-@keyframes graph_shift {
-  0% {
-    transform: translate(0, 0)
-
-  }
-  100% {
-    transform: translate(25px, -30px);
-  }
-}
-
-@media (min-aspect-ratio: 1/1) {
-  .chart_info_card {
-    position: fixed;
-    font-family: Inter;
-    width: 27rem;
-    top: calc(50% + 180px);
-    right: calc(50% + 180px);
-    font-weight: 200;
-    line-height: 1rem;
-    padding: 0;
-    margin: 0;
-  }
-}
-
-@media (max-aspect-ratio: 1/1) {
-  .chart_info_card {
-    position: fixed;
-    font-family: Inter;
-    width: 90vw;
-    bottom: 5vh;
-    left: 5vw;
-    font-weight: 200;
-    line-height: 1rem;
-    padding: 0;
-    margin: 0;
-  }
-
-  .chart_info_card:hover,
-  .chart_info_card:focus
-  .chart_info_card:active {
-    position: fixed;
-    font-family: Inter;
-    width: 90vw;
-    bottom: 5vh;
-    left: 5vw;
-    font-weight: 200;
-    line-height: 1rem;
-    padding: 0;
-    margin: 0;
-    transform: translateX(90vw);
-  }
-}
-
-@keyframes card_load {
-  0% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 1;
-  }
-}
-
-.card_title, .card_year {
-  display: inline;
-  font-size: 1.5rem;
-}
-
-.card_title {
-  font-weight: 600;
-
-  opacity: 0;
-  animation: card_load 5s ease-in;
-  animation-delay: 12s;
-  animation-fill-mode: forwards;
-}
-
-.card_year {
-  opacity: 0;
-  animation: card_load 5s ease-in;
-  animation-delay: 12s;
-  animation-fill-mode: forwards;
-}
-
-.card_medium {
-  margin-top: 0.9rem;
-  font-size: 1.1rem;
-
-  opacity: 0;
-  animation: card_load 5s ease-in;
-  animation-delay: 17s;
-  animation-fill-mode: forwards;
-}
-
-.card_description {
-  margin-top: 0.7rem;
-  font-size: 0.9rem;
-
-  opacity: 0;
-  animation: card_load 5s ease-in;
-  animation-delay: 22s;
-  animation-fill-mode: forwards;
-}
-
-.card_instructions {
-  margin-top: 2rem;
-  font-size: 0.9rem;
-  font-weight: 500;
-
-  opacity: 0;
-  animation: card_load 5s ease-in;
-  animation-delay: 37s;
-  animation-fill-mode: forwards;
-}
-
-.code {
-  background: #fafafa;
-  border-radius: 5px;
-  padding: 0.75rem;
-  font-size: 1.1rem;
-  font-family: Menlo, Monaco, Lucida Console, Liberation Mono, DejaVu Sans Mono,
-    Bitstream Vera Sans Mono, Courier New, monospace;
-}
-
-.grid {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  flex-wrap: wrap;
-  max-width: 800px;
-}
-
-.card {
-  margin: 1rem;
-  padding: 1.5rem;
-  text-align: left;
-  color: inherit;
-  text-decoration: none;
-  border: 1px solid #eaeaea;
-  border-radius: 10px;
-  transition: color 0.15s ease, border-color 0.15s ease;
-  max-width: 300px;
-}
-
-.card:hover,
-.card:focus,
-.card:active {
-  color: #0070f3;
-  border-color: #0070f3;
-}
-
-.card h2 {
-  margin: 0 0 1rem 0;
-  font-size: 1.5rem;
-}
-
-.card p {
-  margin: 0;
-  font-size: 1.25rem;
-  line-height: 1.5;
-}
-
-.logo {
-  height: 1em;
-  margin-left: 0.5rem;
-}
-
-@media (max-width: 600px) {
-  .grid {
-    width: 100%;
-    flex-direction: column;
-  }
-}
-
-@media (prefers-color-scheme: light) {
-  .card,
-  .footer {
-    border-color: #222;
-  }
-  .code {
-    background: #111;
-  }
-  .logo img {
-    filter: invert(1);
-  }
-  .description {
-    color: white;
-  }
-}
-
-@media (prefers-color-scheme: dark) {
-  .card,
-  .footer {
-    border-color: #222;
-  }
-  .code {
-    background: #111;
-  }
-  .logo img {
-    filter: invert(1);
-  }
-  .description {
-    color: white;
-  }
-}
-
