+++ styles/header.module.css
@@ -1,0 +1,32 @@
+.homeLink {
+  color: inherit;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  flex-grow: 1;
+}
+
+.title_container {
+  padding: 0;
+  margin: 0;
+  flex: 1;
+  display: flex;
+  flex-direction: column;
+  position: fixed;
+  top: 10px;
+  left: 10px;
+  z-index: 99;
+  cursor: pointer;
+}
+
+.logo_block {
+  display: flex;
+  flex-direction: row;
+  z-index: 99;
+}
+
+.logo_block img {
+  margin-top: 5px;
+  margin-right: 5px;
+  background-color: white;
+}
