+++ styles/living_reduction.module.css
@@ -1,0 +1,140 @@
+.container {
+  padding: 0 0rem;
+}
+
+.main {
+  min-height: 100vh;
+  width: 100%;
+  padding: 0;
+  margin: 0;
+  flex: 1;
+  position: absolute;
+  top: 0;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+}
+
+.character_tree {
+  opacity: 1;
+  animation: graph_shift;
+  animation-duration: 5s;
+  animation-timing-function: ease-in-out;
+  animation-delay: 10s;
+  animation-fill-mode: forwards;
+}
+
+@keyframes graph_shift {
+  0% {
+    transform: translate(0, 0)
+
+  }
+  100% {
+    transform: translate(25px, -30px);
+  }
+}
+
+@media (min-aspect-ratio: 1/1) {
+  .chart_info_card {
+    position: fixed;
+    font-family: Inter;
+    width: 27rem;
+    top: calc(50% + 180px);
+    right: calc(50% + 180px);
+    font-weight: 200;
+    line-height: 1rem;
+    padding: 0;
+    margin: 0;
+  }
+}
+
+@media (max-aspect-ratio: 1/1) {
+  .chart_info_card {
+    position: fixed;
+    font-family: Inter;
+    width: 90vw;
+    bottom: 5vh;
+    left: 5vw;
+    font-weight: 200;
+    line-height: 1rem;
+    padding: 0;
+    margin: 0;
+  }
+
+  .chart_info_card:hover,
+  .chart_info_card:focus
+  .chart_info_card:active {
+    position: fixed;
+    font-family: Inter;
+    width: 90vw;
+    bottom: 5vh;
+    left: 5vw;
+    font-weight: 200;
+    line-height: 1rem;
+    padding: 0;
+    margin: 0;
+    transform: translateX(90vw);
+  }
+}
+
+@keyframes card_load {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+
+.card_title, .card_year {
+  display: inline;
+  font-size: 1.5rem;
+}
+
+.card_title {
+  font-weight: 600;
+
+  opacity: 0;
+  animation: card_load 5s ease-in;
+  animation-delay: 12s;
+  animation-fill-mode: forwards;
+}
+
+.card_year {
+  opacity: 0;
+  animation: card_load 5s ease-in;
+  animation-delay: 12s;
+  animation-fill-mode: forwards;
+}
+
+.card_medium {
+  margin-top: 0.9rem;
+  font-size: 1.1rem;
+
+  opacity: 0;
+  animation: card_load 5s ease-in;
+  animation-delay: 14.5s;
+  animation-fill-mode: forwards;
+}
+
+.card_description {
+  margin-top: 0.7rem;
+  font-size: 0.9rem;
+
+  opacity: 0;
+  animation: card_load 5s ease-in;
+  animation-delay: 17s;
+  animation-fill-mode: forwards;
+}
+
+.card_instructions {
+  margin-top: 2rem;
+  font-size: 0.9rem;
+  font-weight: 500;
+
+  opacity: 0;
+  animation: card_load 5s ease-in;
+  animation-delay: 22s;
+  animation-fill-mode: forwards;
+}
